+++ appkit-ut/DelimitedTxtSuite.hpp
@@ -81,7 +81,7 @@ private:
     static void cb4a(void*, const char*, size_t);
     static void cb4b(void*, const char*, size_t);
 
-    static appkit::String formExpectedLine(unsigned long);
+    static appkit::String formExpectedLine(size_t);
 
 };
